+++ Publish_Weaver/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.5.1
 * Fix sorter taking from storage to assembler would ignore assembler item limit and put items into assembler forever.
+* Fix advanced mining machine belts input/output would not work when the player left the planet.
 
 ## 1.5.0
 * 10-20% DSP performance improvement compared to 1.4.2